remove sorting from character.ex
The fixture character I was creating was creating abilities in
regular order (str-cha), but when that is saved in the database
it is saved in the reverse order.  Since the form will only ever
save them in the correct order, the change was to update the
character generation in the test to generate the character as
the form would (ie order the abilities cha-str).  Once I did that,
the abilities appeared in the correct order in the tests, thus
removing the need to sort the fields in the changeset function
(which prevented saving of the form unless all that character's
fields appeared on the form at once).